requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.8.0
dm-tree==0.1.8
dnspython==2.4.2
exceptiongroup==1.2.0
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.10.0
gast==0.4.0
google-auth==2.24.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.9.0
jax==0.4.20
Jinja2==3.1.2
joblib==1.3.2
keras==2.12.0
keras-nightly==3.0.1.dev2023120203
libclang==16.0.6
loguru==0.7.2
mando==0.7.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.3.1
more-itertools==10.1.0
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
openai==1.3.5
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pinecone-client==2.2.4
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
radon==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sympy==1.12
tb-nightly==2.16.0a20231202
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.12.0
termcolor==2.3.0
tf-estimator-nightly==2.14.0.dev2023080308
tf-keras-nightly==2.16.0.dev2023120210
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
